// components/UdyamMap.d.ts
import type React from "react";

export declare const MapView: React.ComponentType<any>;
export declare const Marker: React.ComponentType<any>;
export declare const Circle: React.ComponentType<any>;
export declare const PROVIDER_GOOGLE: any;

export type Region = {
  latitude: number;
  longitude: number;
  latitudeDelta: number;
  longitudeDelta: number;
};